fixing bugs in sessionpreset entities, missing newly added fields...

+++ screens/src/main/java/fr/shining_cat/everyday/screens/viewmodels/sessionpresets/AbstractSessionPresetViewModel.kt
@@ -77,7 +77,7 @@ abstract class AbstractSessionPresetViewModel(
 
     abstract fun updatePresetVibration(inputVibration: Boolean)
 
-    abstract fun updatePresetSessionTypeId(inputSessionTypeId: Int)
+    abstract fun updatePresetSessionTypeId(inputSessionTypeId: Long)
 
     fun saveSessionPreset(sessionPreset: SessionPreset) {
         viewModelScope.launch {
